chore: remove nango frontend (#523)
## Description

This removes the use of nango frontend

## Additional Notes

Add any other context or screenshots about the feature request here.

## Checklist:

Before you create this PR, confirm all the requirements listed below by
checking the checkboxes like this (`[x]`).

- [ ] I have performed a self-review of my own code.
- [ ] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have added tests to cover the new feature or fixes.
