+++ packages/nango/package.json
@@ -8,7 +8,6 @@
     "type-check": "tsc --noEmit"
   },
   "dependencies": {
-    "@nangohq/frontend": "^0.42.20",
     "@nangohq/types": "^0.42.20"
   },
   "devDependencies": {
