+++ packages/nango/src/api/error.ts
@@ -1,0 +1,11 @@
+export class NangoConnectionError extends Error {
+  readonly response: Response;
+  constructor(
+    message: string,
+    { response, ...opts }: Parameters<ErrorConstructor>[1] & { response: Response }
+  ) {
+    super(message, opts);
+    this.name = 'NangoConnectionError';
+    this.response = response;
+  }
+}
