+++ packages/nango/src/index.ts
@@ -1,2 +1,1 @@
-export * from './frontend';
 export * from './api';
